cleaned up code + fix problems
Made different files to clean up main.dart and fixed problems provided by dart to make our app faster/smoother.
